requirements.txt
fastapi
uvicorn
aioslsk
python-decouple